feture UI setBrightnessWithColor

+++ str/apps/include/EtRobocon2018.h
@@ -38,9 +38,9 @@ class EtRobocon2018 {
 
  private:
   TouchSensor touchSensor;
-  int8_t light_white;
-  int8_t light_black;
-  int8_t target_brightness;
+  int16_t light_white;
+  int16_t light_black;
+  int16_t target_brightness;
   int32_t firstCode;
   ColorSensor colorSensor;
   UserInterface ui;
